# backend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
    "/api/simulation/initialize" : {
      "post" : {
        "tags" : [ "simulation-controller" ],
        "operationId" : "initializeSimulation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimulationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimulationResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/ground-truth" : {
      "get" : {
        "tags" : [ "simulation-controller" ],
        "operationId" : "getGroundTruth",
        "parameters" : [ {
          "name" : "body",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "frame",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "fromEpoch",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "toEpoch",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "stepSeconds",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "subtractSun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroundTruthResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "simulation-controller" ],
        "operationId" : "sampleGroundTruth",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroundTruthSampleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroundTruthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/chunk" : {
      "post" : {
        "tags" : [ "simulation-controller" ],
        "operationId" : "getNextChunk",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimulationChunkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "SimulationRequestDTO" : {
        "type" : "object",
        "properties" : {
          "celestialBodyNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "date" : {
            "type" : "string"
          },
          "frame" : {
            "type" : "string"
          },
          "integrator" : {
            "type" : "string"
          },
          "timeStepUnit" : {
            "type" : "string"
          },
          "fidelityBucket" : {
            "type" : "string"
          },
          "previousSessionID" : {
            "type" : "string"
          }
        }
      },
      "CelestialBodyWrapper" : {
        "type" : "object",
        "properties" : {
          "mass" : {
            "type" : "number",
            "format" : "double"
          },
          "mu" : {
            "type" : "number",
            "format" : "double"
          },
          "radius" : {
            "type" : "number",
            "format" : "double"
          },
          "name" : {
            "type" : "string"
          },
          "orbitingBody" : {
            "type" : "string"
          },
          "referenceBodyNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "SimulationResponseDTO" : {
        "type" : "object",
        "properties" : {
          "celestialBodyPropertiesList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CelestialBodyWrapper"
            }
          },
          "simulationMetaData" : {
            "$ref" : "#/components/schemas/SimulationResponseMetadata"
          }
        }
      },
      "SimulationResponseMetadata" : {
        "type" : "object",
        "properties" : {
          "sessionID" : {
            "type" : "string"
          }
        }
      },
      "GroundTruthSampleRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "frame" : {
            "type" : "string"
          },
          "referenceEpochs" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "subtractSun" : {
            "type" : "boolean"
          }
        }
      },
      "BodyGroundTruthTrack" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "anchors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GroundTruthAnchor"
            }
          }
        }
      },
      "GroundTruthAnchor" : {
        "type" : "object",
        "properties" : {
          "epochMillis" : {
            "type" : "integer",
            "format" : "int64"
          },
          "referenceEpoch" : {
            "type" : "integer",
            "format" : "int64"
          },
          "position" : {
            "type" : "array",
            "items" : {
              "type" : "number",
              "format" : "double"
            }
          },
          "velocity" : {
            "type" : "array",
            "items" : {
              "type" : "number",
              "format" : "double"
            }
          }
        }
      },
      "GroundTruthResponse" : {
        "type" : "object",
        "properties" : {
          "tracks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BodyGroundTruthTrack"
            }
          }
        }
      },
      "SimulationChunkRequest" : {
        "type" : "object",
        "properties" : {
          "sessionID" : {
            "type" : "string"
          },
          "expectedChunkIndex" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}